Add dataclass for visual effects
A dataclass is added to replace the logic_data dict. logic_data has
complex typing and requires various type checks when
iterating over. In addition, the keys were hardcoded to colors
reducing options.

The dataclass Effect has specific variables for different types and
a color variable as well as a layer variable for prioritizing multiple
effects on the same cell.